+++ cmd/simplelog/logger.go
@@ -21,6 +21,7 @@ import (
 	"log"
 	"os"
 	"path"
+	"sync"
 )
 
 const (
@@ -30,8 +31,9 @@ const (
 	LevelDebug
 )
 
-var LOGGER *Logger
+var logger *Logger
 var ddcLog *os.File
+var mut = &sync.Mutex{}
 
 type Logger struct {
 	debugLogger   *log.Logger
@@ -41,34 +43,44 @@ type Logger struct {
 }
 
 func init() {
-	LOGGER = NewLogger(LevelError)
+	logger = newLogger(LevelError)
 }
 
 func InitLogger(level int) {
 	if level > 3 {
-		LOGGER = NewLogger(LevelDebug)
+		logger = newLogger(LevelDebug)
 	} else {
-		LOGGER = NewLogger(level)
+		logger = newLogger(level)
 	}
 }
 
 func Close() error {
+	logger.Info("Close called on log")
 	if err := ddcLog.Close(); err != nil {
 		return fmt.Errorf("unable to close ddc.log with error %v", err)
 	}
 	return nil
 }
 
-func NewLogger(level int) *Logger {
+func newLogger(level int) *Logger {
 	debugOut, infoOut, warningOut, errorOut := io.Discard, io.Discard, io.Discard, io.Discard
 	ddcLoc, err := os.Executable()
 	if err != nil {
 		log.Fatalf("unable to to find ddc cannot copy it to hosts due to error '%v'", err)
 	}
+	mut.Lock()
+	if ddcLog != nil {
+		log.Printf("closing log")
+		if err := Close(); err != nil {
+			log.Printf("WARN: unable to close log %v", err)
+		}
+	}
+	log.Print("initialized log")
 	ddcLog, err = os.OpenFile(path.Join(path.Dir(ddcLoc), "ddc.log"), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
 	if err != nil {
 		log.Fatal(err)
 	}
+	mut.Unlock()
 	adjustedLevel := level
 	if adjustedLevel > 3 {
 		adjustedLevel = 3
@@ -96,73 +108,81 @@ func NewLogger(level int) *Logger {
 }
 
 func (l *Logger) Debug(format string) {
-	handleLogError(l.debugLogger.Output(2, format))
+	handleLogError(l.debugLogger.Output(2, format), format, "DEBUG")
 }
 
 func (l *Logger) Info(format string) {
-	handleLogError(l.infoLogger.Output(2, format))
+	handleLogError(l.infoLogger.Output(2, format), format, "INFO")
 }
 
 func (l *Logger) Warning(format string) {
-	handleLogError(l.warningLogger.Output(2, format))
+	handleLogError(l.warningLogger.Output(2, format), format, "WARNING")
 }
 
 func (l *Logger) Error(format string) {
-	handleLogError(l.errorLogger.Output(2, format))
+	handleLogError(l.errorLogger.Output(2, format), format, "ERROR")
 }
 
 func (l *Logger) Debugf(format string, v ...interface{}) {
-	handleLogError(l.debugLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(l.debugLogger.Output(2, msg), msg, "DEBUGF")
 }
 
 func (l *Logger) Infof(format string, v ...interface{}) {
-	handleLogError(l.infoLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(l.infoLogger.Output(2, msg), msg, "INFOF")
 }
 
 func (l *Logger) Warningf(format string, v ...interface{}) {
-	handleLogError(l.warningLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(l.warningLogger.Output(2, msg), msg, "WARNINGF")
 }
 
 func (l *Logger) Errorf(format string, v ...interface{}) {
-	handleLogError(l.errorLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(l.errorLogger.Output(2, msg), msg, "ERRORF")
 }
 
 // package functions
 
 func Debug(format string) {
-	handleLogError(LOGGER.debugLogger.Output(2, format))
+	handleLogError(logger.debugLogger.Output(2, format), format, "DEBUG")
 }
 
 func Info(format string) {
-	handleLogError(LOGGER.infoLogger.Output(2, format))
+	handleLogError(logger.infoLogger.Output(2, format), format, "INFO")
 }
 
 func Warning(format string) {
-	handleLogError(LOGGER.warningLogger.Output(2, format))
+	handleLogError(logger.warningLogger.Output(2, format), format, "WARNING")
 }
 
 func Error(format string) {
-	handleLogError(LOGGER.errorLogger.Output(2, format))
+	handleLogError(logger.errorLogger.Output(2, format), format, "ERROR")
 }
 
 func Debugf(format string, v ...interface{}) {
-	handleLogError(LOGGER.debugLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(logger.debugLogger.Output(2, msg), msg, "DEBUGF")
 }
 
 func Infof(format string, v ...interface{}) {
-	handleLogError(LOGGER.infoLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(logger.infoLogger.Output(2, msg), msg, "INFOF")
 }
 
 func Warningf(format string, v ...interface{}) {
-	handleLogError(LOGGER.warningLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(logger.warningLogger.Output(2, msg), msg, "WARNINGF")
 }
 
 func Errorf(format string, v ...interface{}) {
-	handleLogError(LOGGER.errorLogger.Output(2, fmt.Sprintf(format, v...)))
+	msg := fmt.Sprintf(format, v...)
+	handleLogError(logger.errorLogger.Output(2, msg), msg, "ERRORF")
 }
 
-func handleLogError(err error) {
+func handleLogError(err error, attemptedMsg, level string) {
 	if err != nil {
-		log.Fatalf("critical error and therefore there is no log output due to error '%v'", err)
+		log.Fatalf("critical error logging to level %v with message '%v' and therefore there is no log output due to error '%v'", level, attemptedMsg, err)
 	}
 }