+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [0.3.1]
 
+### Added
+
+* support for untrusted certs when accessing the dremio API
+
 ### Fixed
 
 * We were logging WARN messages for unsupported kinds, we are ignoring those now and there will be a debug log that explains we are not masking those kinds
@@ -11,6 +15,7 @@
 
 * if unable to detect any dremio PID process we will now exit with an error that states one must run sudo to run collections
 * several warnings have the text error in them, this was leading to people thinking there was an error when really it is just a warning that may or may not be signficant
+* REST calls now only occur on the first coordinator and will not occur on subsequent nodes
 
 ## [0.3.0]